+++ postgresql/views/03_v_iridium_services_info.sql
@@ -42,3 +42,6 @@ COMMENT ON VIEW v_iridium_services_info IS 'View для доступа к дан
 
 
 
+
+
+